# LibGfx/WebPWriter: Implement run-length encoding